% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Env_shift}
\alias{Env_shift}
\title{Compute a white noise environment with a shift}
\usage{
Env_shift(env, t, rho, env.args)
}
\arguments{
\item{env}{previous year's environment NOT used in current formulation}

\item{t}{the time point}

\item{rho}{the correlation between environment of selection and the cue}

\item{env.args}{other args:
t.jump the location for the jump,
delta env change that takes place at t.jump
sd noise in env
sdc noise in cue}
}
\description{
Compute a white noise environment with a shift
}
\details{
nada
}

