% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Env_shift_cpp}
\alias{Env_shift_cpp}
\title{Compute a white noise environment with a shift}
\usage{
Env_shift_cpp(t, env_args)
}
\arguments{
\item{t}{the time point}

\item{env_args}{other args:
t.jump the location for the jump,
delta env change that takes place at t.jump
sd noise in env
sdc noise in cue}
}
\description{
Compute a white noise environment with a shift
}
\details{
nada
}

