% Generated by roxygen2 (4.0.2): do not edit by hand
\name{G}
\alias{G}
\title{G matrix}
\usage{
G(Gaa, Gbb, Gab)
}
\arguments{
\item{Gaa}{additive genetic variance in RN intercept}

\item{Gbb}{additive genetic variance in RN slope}

\item{Gab}{additive genetic covariance between RN slope and height}
}
\description{
G matrix
}

