% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Log_R_bar}
\alias{Log_R_bar}
\title{Compute LOG population growth rate under stabilizing selection}
\usage{
Log_R_bar(R0, logWbar)
}
\arguments{
\item{R0}{basic reproductive number}

\item{logWbar}{log average fitness}
}
\description{
Compute LOG population growth rate under stabilizing selection
}
\details{
Assumes theta-logistic population regulation
would be good to have separate DD function specified after
chevin and lande 2010
}

