% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Log_W_bar}
\alias{Log_W_bar}
\title{Compute mean fitness under stabilizing selection}
\usage{
Log_W_bar(zbar, theta, oz2, gamma)
}
\arguments{
\item{zbar}{mean trait prior to selection}

\item{theta}{enironmental optimum}

\item{oz2}{strength of stabilizing selection}

\item{gamma}{1/(oz2 + Vz2), with Vz2 phenotypic variance}
}
\description{
Compute mean fitness under stabilizing selection
}
\details{
IN USE
}

