% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_env}
\alias{make_env}
\title{Compute correctly-scaled environment of development and selection}
\usage{
make_env(T, env_args)
}
\arguments{
\item{T}{how many gens}

\item{env_args}{environment args}
}
\description{
Compute correctly-scaled environment of development and selection
}
\details{
given a pre-generated vector of random environments
}

