% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pdRespVarLoad}
\alias{pdRespVarLoad}
\title{Compute change in fitness over one generation after stabilizing selection
function of environment (after Lande Chevin)}
\usage{
pdRespVarLoad(t, X, params, G, env_args)
}
\arguments{
\item{t}{timestep}

\item{X}{parameters (a, b, env, logN)}

\item{params}{a list with (Oz2, AB, R0, K, theta)}

\item{G}{the (constant) G matrix}

\item{env_args}{extra args for env.fn}
}
\value{
rowvector: log R bar, delta R bar, variance load
}
\description{
Compute change in fitness over one generation after stabilizing selection
function of environment (after Lande Chevin)
}
\details{
function signature as for use with deSolve
imposes fitness based on Lande
and demography after CL 2010
}

