% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Pheno_demo}
\alias{Pheno_demo}
\title{Compute trait + demographic change under stabilizing selection as function of environment}
\usage{
Pheno_demo(t, X, p, G, env.fn, env.args)
}
\arguments{
\item{t}{timestep}

\item{X}{parameters (a, b, env, N)}

\item{p}{(rho, gamma, A, B, R0)}

\item{G}{the (constant) G matrix}

\item{env.fn}{function to compute environment with signature function(environment, time, ...)}

\item{env.args}{extra args for env.fn}
}
\description{
Compute trait + demographic change under stabilizing selection as function of environment
}
\details{
function signature for use with deSolve and 'iterate', imposes fitness based on
assumed tolerance curve
}

