% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Pheno_demo_econ}
\alias{Pheno_demo_econ}
\title{Compute trait + demographic change under stabilizing selection as function of environment via Lande Chevin
but with intervention}
\usage{
Pheno_demo_econ(t, X, p, G, env.fn, env.args, interv)
}
\arguments{
\item{t}{timestep}

\item{X}{parameters (a, b, env, N)}

\item{p}{(rho, gamma, A, B, R0, K, thetaL)}

\item{G}{the (constant) G matrix}

\item{env.fn}{function to compute environment with signature function(environment, time, ...)}

\item{env.args}{extra args for env.fn}

\item{interv}{a function of the time, policy to modify the cue}
}
\description{
Compute trait + demographic change under stabilizing selection as function of environment via Lande Chevin
but with intervention
}
\details{
function signature for use with deSolve and 'iterate', imposes fitness based on Lande
and demography after CL 2010
}

