% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Pheno_demo_lande}
\alias{Pheno_demo_lande}
\title{Compute trait + demographic change under stabilizing selection as function of environment via Lande Chevin}
\usage{
Pheno_demo_lande(t, X, p, G, env.fn, env.args)
}
\arguments{
\item{t}{timestep}

\item{X}{parameters (a, b, env, N)}

\item{p}{(rho, gamma, A, B, R0, K, thetaL)}

\item{G}{the (constant) G matrix}

\item{env.fn}{function to compute environment with signature function(environment, time, ...)}

\item{env.args}{extra args for env.fn}
}
\description{
Compute trait + demographic change under stabilizing selection as function of environment via Lande Chevin
}
\details{
function signature for use with deSolve and 'iterate', imposes fitness based on Lande
and demography after CL 2010
}

