% Generated by roxygen2 (4.0.2): do not edit by hand
\name{R_bar_ceiling}
\alias{R_bar_ceiling}
\title{Compute population growth rate under stabilizing selection and ceiling regulation}
\usage{
R_bar_ceiling(R0, Wbar, N, K)
}
\arguments{
\item{R0}{basic reproductive number}

\item{Wbar}{average fitness}

\item{N}{number of individuals in this generation}

\item{K}{carrying capacity}
}
\description{
Compute population growth rate under stabilizing selection and ceiling regulation
}
\details{
Assumes ceiling population regulation
would be good to have separate DD function specified after
chevin and lande 2010
}

