% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{R_bar_gompertz}
\alias{R_bar_gompertz}
\title{Compute population growth rate under stabilizing selection and Gompertz regulation}
\usage{
R_bar_gompertz(R0, Wbar, N, K0)
}
\arguments{
\item{R0}{basic reproductive number}

\item{Wbar}{average fitness}

\item{N}{number of individuals in this generation}

\item{K0}{carrying capacity at optimum trait}
}
\description{
Compute population growth rate under stabilizing selection and Gompertz regulation
}
\details{
Assumes Gompertz population regulation
}

