% Generated by roxygen2 (4.0.2): do not edit by hand
\name{R_bar_thetalog}
\alias{R_bar_thetalog}
\title{Compute population growth rate under stabilizing selection and theta-logistic regulation}
\usage{
R_bar_thetalog(R0, Wbar, N, K, thetaL)
}
\arguments{
\item{R0}{basic reproductive number}

\item{Wbar}{average fitness}

\item{N}{number of individuals in this generation}

\item{K}{carrying capacity}

\item{thetaL}{theta-logistic parameter for density dependence}
}
\description{
Compute population growth rate under stabilizing selection and theta-logistic regulation
}
\details{
Assumes theta-logistic population regulation
would be good to have separate DD function specified after
chevin and lande 2010
}

