% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims.R
\name{run_sims_on_grid}
\alias{run_sims_on_grid}
\title{run a simulation on a parameter grid and summarize it}
\usage{
run_sims_on_grid(summary_function, d, Tlim, ...)
}
\arguments{
\item{summary_function}{a function that returns a data frame based on
simulation output}

\item{d}{a data.frame that defines the grid of parameters to run the
simulation on}

\item{Tlim}{the length of time to run}

\item{...}{should include params for simulate_timeseries}
}
\description{
run a simulation on a parameter grid and summarize it
}
\details{
d - data frame with columns of rho, delta, omegaz, alpha, Vb
value a data.frame with possibly many reps per parameter or some summary
statistic. actual value depends on return value of summary_function.
built-in summary functions include the identity (no summary) `summary_id`
}

