% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SHC}
\alias{SHC}
\title{SHC}
\usage{
SHC(sigma_g2, omegaz2, N, R0)
}
\arguments{
\item{sigma_g2}{asymptotic genetic variance}

\item{omegaz2}{defines width of the fitness function $\omega ^2 + 1$}

\item{N}{census population size}

\item{R0}{reproductive rate}
}
\description{
SHC
}
\details{
For this modified SHC after Burger
         \[
         \sigma_{SHC}(\sigma_g^2) =
          \frac{sigma_g}{1 + \frac{V_s}{\alpha ^ 2 N_e}}
         \]
         For $\alpha^2$ the variance of the effect of new mutations, $V_m$
         the genomic mutation rate per generation, and $N_e \approx 2 R_0 N  / (2 R_0 - 1)$.
         As population size increases, the function approaches a limit of $\sigma_g^2$
         $V_s = \omega^2 + \sigma_e^2$ is the width of the fitness function.
         Hard-coded parameters and $alpha^2 = 0.005$ and $\sigma_e^2 = 1$
}

