\name{stab_curve}
\alias{stab_curve}
\title{Mean fitness -- with constraint, need to make as function of environment...}
\usage{
  stab_curve(x, LL)
}
\arguments{
  \item{x}{the environment}

  \item{LL}{lethal limit (symmetric)}
}
\description{
  Mean fitness -- with constraint, need to make as function
  of environment...
}
\details{
  fitness surf infinitesimal optimum need to revisit
}

