% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Va}
\alias{Va}
\title{Va additive genetic variance in the phenotype as a function of the environment}
\usage{
Va(env, GG)
}
\arguments{
\item{env}{environment in which plastiicty is cued}

\item{GG}{additive genetic variance-covariance matrix}
}
\description{
Va additive genetic variance in the phenotype as a function of the environment
}

