% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims.R
\name{Vz_}
\alias{Vz_}
\alias{variance_load}
\title{variance load}
\usage{
Vz_(siga2, sigb2, delta, sige2)
}
\arguments{
\item{siga2}{squared genetic variance in reaction norm intercept}

\item{sigb2}{squared genetic variance in reaction norm slope}

\item{delta}{shift in the mean environment}

\item{sige2}{squared microenvronmental variance}
}
\description{
variance load
}

