\name{W_bar}
\alias{W_bar}
\title{Compute log mean fitness under stabilizing selection}
\usage{
  W_bar(zbar, theta, Oz2, gamma, log = TRUE)
}
\arguments{
  \item{zbar}{mean trait prior to selection}

  \item{theta}{environmental optimum}

  \item{Oz2}{strength of stabilizing selection}

  \item{gamma}{1/(Oz2 + Vz2), with Vz2 phenotypic variance}

  \item{log}{(=TRUE) whether to return the log of fitness}
}
\description{
  Compute log mean fitness under stabilizing selection
}
\details{
  no details
}

