% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Beta}
\alias{Beta}
\title{Selection on plasticity as as function of environment assuming stabilizing selection}
\usage{
Beta(gamma, A, B, a, b, e_t, e_plast)
}
\arguments{
\item{gamma}{= 1/(Oz2 + Vz2), with Vz2 phenotypic variance}

\item{A}{the environmental optimum RN int}

\item{B}{the environmental optimum RN slope}

\item{a}{current value of RN int}

\item{b}{current value of RN slope}

\item{e_t}{environment now}

\item{e_plast}{env that cues plasticity}
}
\description{
Selection on plasticity as as function of environment assuming stabilizing selection
}

