% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims.R
\name{G}
\alias{G}
\alias{G_matrix}
\title{G matrix}
\usage{
G(Gaa, Gbb, Gab)
}
\arguments{
\item{Gaa}{additive genetic variance in RN intercept}

\item{Gbb}{additive genetic variance in RN slope}

\item{Gab}{additive genetic covariance between RN slope and height}
}
\description{
G matrix
}

