% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R_bar}
\alias{R_bar}
\title{Compute population growth rate under stabilizing selection}
\usage{
R_bar(R0, Wbar)
}
\arguments{
\item{R0}{basic reproductive number}

\item{Wbar}{average fitness}
}
\description{
Compute population growth rate under stabilizing selection
}
\details{
Assumes density-independent
}

