% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{R_bar}
\alias{R_bar}
\title{Compute population growth rate under stabilizing selection and no regulation}
\usage{
R_bar(R0, Wbar, N)
}
\arguments{
\item{R0}{basic reproductive number}

\item{Wbar}{average fitness}

\item{N}{number of individuals in this generation}

\item{K}{carrying capacity}
}
\description{
Compute population growth rate under stabilizing selection and no regulation
}
\details{
Assumes ceiling population regulation
would be good to have separate DD function specified after
chevin and lande 2010
}

