% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_pheno_ts}
\alias{simulate_pheno_ts}
\title{Compute phenotypic dynamic Time Series of trait + demographic change under stabilizing selection as
function of environment (after Lande Chevin)}
\usage{
simulate_pheno_ts(T, X, params, env_args, growth_fun = "independent",
  poisson = FALSE)
}
\arguments{
\item{T}{end time, assuming start time of 1}

\item{X}{parameters (z, a, b, wbar, logN, theta)}

\item{params}{a list with (gamma_sh, omegaz, A, B, R0, var_a, Vb, delta, sigma_xi, rho_tau, fractgen)}

\item{env_args}{extra args for env.fn}

\item{growth_fun}{density-dependence ("independent", "gompertz", "thetalogistic", "ceiling")}

\item{poisson}{(logical) return N(t+1) = Poisson(f(N(t))}
}
\value{
a long matrix with columns zbar, abar, bbar, Wbar, Npop, theta
}
\description{
Compute phenotypic dynamic Time Series of trait + demographic change under stabilizing selection as
function of environment (after Lande Chevin)
}
\details{
NB - for now assume Tchange = 0 and demography after CL 2010
         poisson only implemented for default
}

