% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims.R
\name{simulate_timeseries}
\alias{simulate_timeseries}
\title{simulate a timeseries}
\usage{
simulate_timeseries(nrep, Tlim, delta, rho, alpha, omegaz, Vb, sigma_xi, Npop0,
  var_a, Ve = Ve, fractgen = fractgen, omega_Wmax = omega_Wmax, A = A,
  B = B, stationarity = TRUE, ...)
}
\arguments{
\item{nrep}{reps}

\item{Tlim}{max time}

\item{delta}{shift}

\item{rho}{predictability of the environmental noise}

\item{alpha}{intial mean plasticity}

\item{omegaz}{selection}

\item{Vb}{variance in RN slope}

\item{sigma_xi}{variance in the optimum}

\item{Npop0}{initial population}

\item{var_a}{variance in trait}

\item{Ve}{microenvronemntal variance}

\item{fractgen}{developmental delay}

\item{omega_Wmax}{max productivity}

\item{A}{reference environment intercept}

\item{B}{environmental sensitivity of selection}

\item{stationarity}{(logical) generate stationary IC if TRUE}

\item{...}{avoid throwing errors if we pass too many? (bad idea)}

\item{K0}{carrying capacity (not implemented)}
}
\description{
simulate a timeseries
}
\details{
uses `simulate_pheno_ts` under the hood
}

