\name{Va}
\alias{Va}
\title{Va additive genetic variance in the phenotype as a function of the environment}
\usage{
  Va(e.plast, GG)
}
\arguments{
  \item{e.plast}{environment in which plastiicty is cued}

  \item{GG}{additive genetic variance-covariance matrix}
}
\description{
  Va additive genetic variance in the phenotype as a
  function of the environment
}

