% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W_bar}
\alias{W_bar}
\title{Compute mean fitness under stabilizing selection}
\usage{
W_bar(zbar, theta, oz2, gamma, LOG)
}
\arguments{
\item{zbar}{mean trait prior to selection}

\item{theta}{enironmental optimum}

\item{oz2}{strength of stabilizing selection}

\item{gamma}{1/(oz2 + Vz2), with Vz2 phenotypic variance}

\item{LOG}{(=TRUE) whether to return the log of fitness}
}
\description{
Compute mean fitness under stabilizing selection
}
\details{
IN USE
}

